+++ lib/src/apollovm_parser.dart
@@ -9,14 +9,33 @@ import 'apollovm_base.dart';
 import 'ast/apollovm_ast_toplevel.dart';
 
 /// Base class for [ApolloVM] parsers.
-abstract class ApolloParser {
+abstract class ApolloCodeParser<T extends Object> {
+  /// The language of this parser.
+  String get language;
+
+  /// Parses a [codeUnit] to an [ASTRoot] and returns a [ParseResult].
+  ///
+  /// If some error occurs, returns a [ParseResult] with an error message.
+  Future<ParseResult<T>> parse(CodeUnit<T> codeUnit);
+
+  void check(CodeUnit codeUnit) {
+    if (!acceptsLanguage(codeUnit.language)) {
+      throw StateError(
+          "This parser is for the language '$language'. Trying to parse a CodeUnit of language: '${codeUnit.language}'");
+    }
+  }
+
+  bool acceptsLanguage(String language) {
+    return this.language == language;
+  }
+}
+
+/// Base class for [ApolloVM] source code parsers.
+abstract class ApolloSourceCodeParser extends ApolloCodeParser<String> {
   /// The [GrammarDefinition] of this parser.
   final GrammarDefinition _grammar;
 
-  ApolloParser(this._grammar);
-
-  /// The language of this parser.
-  String get language;
+  ApolloSourceCodeParser(this._grammar);
 
   Parser<dynamic>? _grammarParserInstance;
 
@@ -28,10 +47,11 @@ abstract class ApolloParser {
   /// Parses a [codeUnit] to an [ASTRoot] and returns a [ParseResult].
   ///
   /// If some error occurs, returns a [ParseResult] with an error message.
-  Future<ParseResult> parse(CodeUnit codeUnit) async {
+  @override
+  Future<ParseResult<String>> parse(CodeUnit<String> codeUnit) async {
     check(codeUnit);
 
-    var result = _grammarParser.parse(codeUnit.source);
+    var result = _grammarParser.parse(codeUnit.code);
 
     if (result is! Success) {
       var lineAndColumn = result
@@ -49,25 +69,14 @@ abstract class ApolloParser {
     var root = result.value;
     return ParseResult(codeUnit, root: root);
   }
-
-  void check(CodeUnit codeUnit) {
-    if (!acceptsLanguage(codeUnit.language)) {
-      throw StateError(
-          "This parser is for the language '$language'. Trying to parse a CodeUnit of language: '${codeUnit.language}'");
-    }
-  }
-
-  bool acceptsLanguage(String language) {
-    return this.language == language;
-  }
 }
 
-class ParseResult {
+class ParseResult<T> {
   /// The parsed code.
-  final CodeUnit codeUnit;
+  final CodeUnit<T> codeUnit;
 
-  /// The parsed [codeUnit] source.
-  String get source => codeUnit.source;
+  /// The parsed [codeUnit] code.
+  T get source => codeUnit.code;
 
   /// A parsed [ASTRoot]
   final ASTRoot? root;
@@ -91,7 +100,11 @@ class ParseResult {
   String? get errorLine {
     var lineAndColumn = errorLineAndColumn;
     if (lineAndColumn != null && lineAndColumn.isNotEmpty) {
-      return codeUnit.getLine(lineAndColumn[0]);
+      final codeUnit = this.codeUnit;
+      if (codeUnit is SourceCodeUnit) {
+        var sourceCodeUnit = codeUnit as SourceCodeUnit;
+        return sourceCodeUnit.getLine(lineAndColumn[0]);
+      }
     }
 
     return null;
@@ -106,6 +119,7 @@ class ParseResult {
   /// Returns the [errorMessage] with the error line information.
   String get errorMessageExtended {
     final errorLine = this.errorLine;
+
     if (errorLine != null && errorLine.isNotEmpty) {
       final errorLineAndColumn = this.errorLineAndColumn;