+++ lib/src/apollovm_runner.dart
@@ -9,11 +9,14 @@ import 'ast/apollovm_ast_toplevel.dart';
 import 'ast/apollovm_ast_type.dart';
 import 'ast/apollovm_ast_value.dart';
 
+@Deprecated("Renamed to `ApolloRunner`")
+typedef ApolloLanguageRunner = ApolloRunner;
+
 /// Base class for [ApolloVM] runners.
 ///
 /// Implementations of this class allows the execution of an [ASTRoot]
 /// in a specific [language].
-abstract class ApolloLanguageRunner implements VMTypeResolver {
+abstract class ApolloRunner implements VMTypeResolver {
   /// The [ApolloVM] of this runner.
   final ApolloVM apolloVM;
 
@@ -24,13 +27,13 @@ abstract class ApolloLanguageRunner implements VMTypeResolver {
 
   ApolloExternalFunctionMapper? externalFunctionMapper;
 
-  ApolloLanguageRunner(this.apolloVM) {
+  ApolloRunner(this.apolloVM) {
     _languageNamespaces = apolloVM.getLanguageNamespaces(language);
     externalFunctionMapper = createDefaultApolloExternalFunctionMapper();
   }
 
   /// Returns a copy of this instance.
-  ApolloLanguageRunner copy();
+  ApolloRunner copy();
 
   /// The default [ApolloExternalFunctionMapper] for this target language runner.
   ///
@@ -106,14 +109,9 @@ abstract class ApolloLanguageRunner implements VMTypeResolver {
         externalFunctionMapper: externalFunctionMapper, typeResolver: this);
   }
 
-  /// Executes a function in [namespace] and with name [functionName].
-  ///
-  /// - [positionalParameters] Positional parameters to pass to the function.
-  /// - [namedParameters] Named parameters to pass to the function.
-  FutureOr<ASTValue> executeFunction(String namespace, String functionName,
-      {List? positionalParameters,
-      Map? namedParameters,
-      bool allowClassMethod = false}) async {
+  FutureOr<({CodeUnit? codeUnit, String? className})> getFunctionCodeUnit(
+      String namespace, String functionName,
+      {bool allowClassMethod = false}) {
     var codeNamespace = _languageNamespaces.get(namespace);
 
     var codeUnit = codeNamespace.getCodeUnitWithFunction(functionName);
@@ -125,14 +123,35 @@ abstract class ApolloLanguageRunner implements VMTypeResolver {
           ?.getClassWithMethod(functionName);
 
       if (classWithMethod != null) {
-        return executeClassMethod(namespace, classWithMethod.name, functionName,
-            positionalParameters: positionalParameters,
-            namedParameters: namedParameters);
+        return (codeUnit: codeUnit, className: classWithMethod.name);
       }
     }
 
+    return (codeUnit: codeUnit, className: null);
+  }
+
+  /// Executes a function in [namespace] and with name [functionName].
+  ///
+  /// - [positionalParameters] Positional parameters to pass to the function.
+  /// - [namedParameters] Named parameters to pass to the function.
+  Future<ASTValue> executeFunction(String namespace, String functionName,
+      {List? positionalParameters,
+      Map? namedParameters,
+      bool allowClassMethod = false}) async {
+    var r = await getFunctionCodeUnit(namespace, functionName,
+        allowClassMethod: allowClassMethod);
+
+    var codeUnit = r.codeUnit;
     if (codeUnit == null) {
-      throw StateError("Can't find function to execute: $functionName");
+      throw StateError(
+          "Can't find function to execute> functionName: $functionName ; language: $language");
+    }
+
+    var className = r.className;
+    if (className != null) {
+      return executeClassMethod(namespace, className, functionName,
+          positionalParameters: positionalParameters,
+          namedParameters: namedParameters);
     }
 
     var result = await codeUnit.root!.execute(
@@ -250,6 +269,6 @@ abstract class ApolloLanguageRunner implements VMTypeResolver {
 
   @override
   String toString() {
-    return 'ApolloLanguageRunner{ language: $language, apolloVM: $apolloVM }';
+    return 'ApolloRunner{ language: $language, apolloVM: $apolloVM }';
   }
 }